% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/when_experimenthub_drops_support.R
\name{human_gene_transcript_sce}
\alias{human_gene_transcript_sce}
\title{When experimenthub drops support}
\usage{
human_gene_transcript_sce()
}
\value{
the included dataset
}
\description{
When experimenthub drops support
}
\examples{
human_gene_transcript_sce()
}
