% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{human_gene_transcript_sce}
\alias{human_gene_transcript_sce}
\title{human_gene_transcript_sce}
\format{
a SingleCellExperiment object
}
\source{
https://www.biorxiv.org/content/10.1101/2023.02.28.530247v1
}
\usage{
human_gene_transcript_sce(metadata=FALSE)
}
\arguments{
\item{metadata}{\code{logical} value indicating whether only
metadata should be returned or if the resource
should be loaded}
}
\value{
a SingleCellExperiment object
}
\description{
Human retinal organoid data from
https://www.biorxiv.org/content/10.1101/2023.02.28.530247v1
}
\examples{
human_gene_transcript_sce()

}
\keyword{datasets}
