% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{data}
\name{human_gene_transcript_sce}
\alias{human_gene_transcript_sce}
\title{human_gene_transcript_sce}
\description{
Human retinal organoid data from
https://www.biorxiv.org/content/10.1101/2023.02.28.530247v1
}
\author{
Kevin Stachelek \email{kevin.stachelek@gmail.com}
}
